+++ dags/openshift_nightlies/tasks/benchmarks/e2e.py
@@ -24,8 +24,8 @@ class E2EBenchmarks():
         # General DAG Configuration
         self.dag = dag
         self.release = release
-        self.exec_config = var_loader.get_executor_config_with_cluster_access(self.release)
-        self.snappy_creds = Variable.get("snappy_creds", deserialize_json=True)
+        self.exec_config = executor.get_executor_config_with_cluster_access(self.release)
+        self.snappy_creds = var_loader.get_secret("snappy_creds", deserialize_json=True)
         
         # Specific Task Configuration
         self.vars = var_loader.build_task_vars(
@@ -78,7 +78,7 @@ class E2EBenchmarks():
         benchmark >> indexer 
 
     def _get_benchmark(self, benchmark):
-        env = {**self.env, **benchmark.get('env', {}), **{"ES_SERVER": var_loader.get_elastic_url()}, **{"KUBEADMIN_PASSWORD": environ.get("KUBEADMIN_PASSWORD", "")}}
+        env = {**self.env, **benchmark.get('env', {}), **{"ES_SERVER": var_loader.get_secret('elasticsearch')}, **{"KUBEADMIN_PASSWORD": environ.get("KUBEADMIN_PASSWORD", "")}}
         return BashOperator(
             task_id=f"{benchmark['name']}",
             depends_on_past=False,
