+++ dags/openshift_nightlies/tasks/index/status.py
@@ -20,7 +20,7 @@ class StatusIndexer():
         # General DAG Configuration
         self.dag = dag
         self.release = release
-        self.exec_config = var_loader.get_executor_config_with_cluster_access(release)
+        self.exec_config = executor.get_executor_config_with_cluster_access(release)
     
         # Specific Task Configuration
         self.vars = var_loader.build_task_vars(release, task="index")
@@ -42,7 +42,7 @@ class StatusIndexer():
     def get_index_task(self):
         env = {
             **self.env, 
-            **{"ES_SERVER": var_loader.get_elastic_url()},
+            **{"ES_SERVER": var_loader.get_secret('elasticsearch')},
             **environ
         }
